indian-law-assistant/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
python-multipart==0.0.6
pydantic==2.5.0
typing-extensions==4.8.0
cors==1.0.1
fastapi-cors==0.0.6
python-json-logger==2.0.7